backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
motor>=3.5.0,<4.0.0
pymongo>=4.8.0,<5.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.1.0,<5.0.0
requests>=2.32.0,<3.0.0
beautifulsoup4>=4.12.3,<5.0.0
lxml>=5.2.2,<6.0.0
psutil>=6.0.0,<7.0.0
selenium>=4.24.0,<5.0.0
email-validator>=2.2.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
